refactor: extract action effect group builders
